USB port security API: move implementation to a separate file
