+++ services/usb/java/com/android/server/usb/hal/port/UsbPortAidl.java
@@ -40,8 +40,6 @@ import android.hardware.usb.AltModeData;
 import android.hardware.usb.AltModeData.DisplayPortAltModeData;
 import android.hardware.usb.DisplayPortAltModePinAssignment;
 import android.hardware.usb.flags.Flags;
-import android.hardware.usb.ext.IUsbExt;
-import android.hardware.usb.ext.PortSecurityState;
 import android.os.Build;
 import android.os.ServiceManager;
 import android.os.IBinder;
@@ -463,52 +461,7 @@ public final class UsbPortAidl implements UsbPortHal {
     public void setPortSecurityState(String portName,
                                      @android.hardware.usb.ext.PortSecurityState int state,
                                      android.os.ResultReceiver callback) {
-        IBinder ext;
-        try {
-            ext = mBinder.getExtension();
-        } catch (RemoteException e) {
-            Slog.e(TAG, "", e);
-            throw new UnsupportedOperationException(e);
-        }
-
-        if (ext == null) {
-            Slog.d(TAG, "setPortSecurityState: no IUsbExt");
-            throw new UnsupportedOperationException();
-        }
-
-        var halCallback = new android.hardware.usb.ext.IPortSecurityStateCallback.Stub() {
-            @Override
-            public void onSetPortSecurityStateCompleted(int status, int arg1, String arg2) {
-                Slog.d(TAG, "onSetPortSecurityStateCompleted, status: " + status);
-                android.os.Bundle b = null;
-                if (arg1 != 0 || arg2 != null) {
-                    b = new android.os.Bundle();
-                    b.putInt("arg1", arg1);
-                    b.putString("arg2", arg2);
-                }
-                callback.send(status, b);
-            }
-
-            @Override
-            public String getInterfaceHash() {
-                return android.hardware.usb.ext.IPortSecurityStateCallback.HASH;
-            }
-
-            @Override
-            public int getInterfaceVersion() {
-                return android.hardware.usb.ext.IPortSecurityStateCallback.VERSION;
-            }
-        };
-
-        Slog.d(TAG, "setPortSecurityState, port: " + portName + ", state " + state);
-
-        IUsbExt usbExt = IUsbExt.Stub.asInterface(ext);
-        try {
-            usbExt.setPortSecurityState(portName, state, halCallback);
-        } catch (RemoteException e) {
-            Slog.e(TAG, "", e);
-            throw new android.os.ParcelableException(e);
-        }
+        UsbPortAidlExt.setPortSecurityState(mBinder, portName, state, callback);
     }
 
     private static class HALCallback extends IUsbCallback.Stub {
